auto_mcs150/requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dotenv==0.9.9
idna==3.10
json5==0.12.1
lxml==6.0.0
numpy==2.3.2
pandas==2.3.1
pillow==11.3.0
probableparsing==0.0.1
pycparser==2.22
pypdf==6.0.0
PyPDFForm==3.4.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-safer==2.0
pytz==2025.2
reportlab==4.4.3
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
usaddress==0.5.16

